fix: prevent server-side env vars from being exposed to client in chat app
Following create-t3-turbo pattern, moved TRPC context creation from separate
context.ts file into trpc.ts and renamed createChatContext to createTRPCContext.

This prevents server-side environment variables (DATABASE_HOST, DATABASE_USERNAME,
DATABASE_PASSWORD) from being exposed to the client-side code when importing
from @api/chat.

Changes:
- Move context creation into api/chat/src/trpc.ts
- Rename createChatContext → createTRPCContext
- Remove separate context.ts file
- Update imports throughout chat app
- Follow exact t3-turbo structure for proper env isolation
